💄 feat(sharedUI): adaptive playback-speed panel in commonMain (drop Android haptics)
